+++ app/Exceptions/Handler.php
@@ -24,7 +24,7 @@ use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Str;
 use Symfony\Component\HttpKernel\Exception\HttpException;
-use BristolSU\UnionCloud\Exception\PermissionDeniedException;
+use Throwable;
 
 class Handler extends ExceptionHandler
 {
@@ -39,8 +39,7 @@ class Handler extends ExceptionHandler
         ModuleInactive::class,
         NotInActivityInstanceException::class,
         ActivityDisabled::class,
-        ModuleInstanceDisabled::class,
-        PermissionDeniedException::class
+        ModuleInstanceDisabled::class
     ];
 
     /**
@@ -56,11 +55,11 @@ class Handler extends ExceptionHandler
     /**
      * Report or log an exception.
      *
-     * @param Exception $exception
+     * @param Throwable $exception
      * @return void
-     * @throws Exception
+     * @throws Throwable
      */
-    public function report(Exception $exception)
+    public function report(Throwable $exception)
     {
         parent::report($exception);
     }
@@ -74,13 +73,13 @@ class Handler extends ExceptionHandler
      * - Log into the default activity instance if possible if $e instanceof NotInActivityInstanceException
      *
      * @param Request $request
-     * @param Exception $exception
+     * @param Throwable $exception
      * @return Response|RedirectResponse|JsonResponse
      * @throws ActivityRequiresGroup
      * @throws ActivityRequiresRole
      * @throws ActivityRequiresUser
      */
-    public function render($request, Exception $exception)
+    public function render($request, Throwable $exception)
     {
         if (!$request->expectsJson()) {
 
@@ -132,10 +131,6 @@ class Handler extends ExceptionHandler
             if($exception instanceof IncorrectLogin) {
                 return response()->json(['message' => $exception->getMessage()], $exception->getStatusCode());
             }
-        }
-
-        if($exception instanceof PermissionDeniedException) {
-            throw new HttpException(403, $exception->getMessage());
         }
 
         // TODO Handle exceptions above if expects json