+++ app/Http/Controllers/Api/ActionInstanceController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use BristolSU\Support\Action\ActionInstance;
-use BristolSU\Support\User\Contracts\UserAuthentication;
+use BristolSU\Support\Authentication\Contracts\Authentication;
 use Illuminate\Http\Request;
 
 /**
@@ -32,7 +32,7 @@ class ActionInstanceController extends Controller
     {
         return ActionInstance::create(array_merge($request->only([
             'name', 'description', 'event', 'action', 'module_instance_id', 'should_queue'
-        ]), ['user_id' => app(UserAuthentication::class)->getUser()->controlId()]));
+        ]), ['user_id' => app(Authentication::class)->getUser()->id()]));
     }
 
     /**