+++ app/Http/Controllers/Api/LogicController.php
@@ -5,10 +5,10 @@ namespace App\Http\Controllers\Api;
 
 
 use App\Http\Controllers\Controller;
+use BristolSU\Support\Authentication\Contracts\Authentication;
 use BristolSU\Support\Filters\FilterInstance;
 use BristolSU\Support\Logic\Contracts\LogicRepository;
 use BristolSU\Support\Logic\Logic;
-use BristolSU\Support\User\Contracts\UserAuthentication;
 use Illuminate\Http\Request;
 
 class LogicController extends Controller
@@ -30,7 +30,7 @@ class LogicController extends Controller
         $logic = $repository->create([
             'name' => $request->input('name'),
             'description' => $request->input('description'),
-            'user_id' => app(UserAuthentication::class)->getUser()->controlId()
+            'user_id' => app(Authentication::class)->getUser()->id()
         ]);
 
         // TODO Refactor below