+++ app/Http/View/Header/CurrentAuthComposer.php
@@ -3,7 +3,6 @@
 namespace App\Http\View\Header;
 
 use BristolSU\Support\Authentication\Contracts\Authentication;
-use BristolSU\ControlDB\Contracts\Models\Role;
 use Illuminate\Contracts\View\View;
 
 class CurrentAuthComposer