+++ app/Providers/AppServiceProvider.php
@@ -11,16 +11,10 @@ use App\Filters\Group\PredefinedFilter as PredefinedGroupFilter;
 use App\Filters\Role\PredefinedFilter as PredefinedRoleFilter;
 use App\Filters\User\PredefinedFilter as PredefinedUserFilter;
 use App\Rules\CustomValidator;
-use App\Support\Audience\LogicAudienceCacher;
-use App\Support\Settings\Setting;
-use App\Support\Settings\SettingRepository;
 use BristolSU\Support\Action\Facade\ActionManager;
 use BristolSU\Support\Filters\Contracts\FilterManager as FilterManagerContract;
-use BristolSU\Support\Logic\Contracts\Audience\LogicAudience;
-use Illuminate\Contracts\Cache\Repository;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\ServiceProvider;
-use Laravel\Passport\Passport;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -31,7 +25,6 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        Passport::withoutCookieSerialization();
         $this->app->call([$this, 'registerFilters']);
 
         Validator::resolver(function($translator, $data, $rules, $messages, $attributes)
@@ -48,8 +41,6 @@ class AppServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        $this->app->bind(SettingRepository::class, Setting::class);
-
         $this->app->bind('auth.password.broker', function ($app) {
             return $app->make('auth.password')->broker();
         });