+++ config/app.php
@@ -180,10 +180,11 @@ return [
         App\Providers\AuthServiceProvider::class,
         App\Providers\EventServiceProvider::class,
         App\Providers\RouteServiceProvider::class,
+        App\Providers\VaporUiServiceProvider::class,
         App\Providers\ViewServiceProvider::class,
         \App\Providers\ControlServiceProvider::class,
-        \App\Providers\AdditionalAttributeProvider::class,
-        \App\Providers\GoogleDriveServiceProvider::class
+        \App\Providers\GoogleDriveServiceProvider::class,
+        \App\Providers\SettingServiceProvider::class
     ],
 
     /*