+++ config/template.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'name' => 'Template',
+    'description' => 'A template module to start from when developing custom modules',
+];