+++ resources/js/settings.js
@@ -1,0 +1,45 @@
+export default {
+    "authentication": {
+        "credentials": {
+            "identifier": "authentication.credentials.identifier"
+        },
+        "access": {
+            "registration-enabled": "authentication.access.registration-enabled",
+            "controluser-registration-enabled": "authentication.access.controluser-registration-enabled",
+            "datauser-registration-enabled": "authentication.access.datauser-registration-enabled",
+            "default-home": "authentication.access.default-home"
+        },
+        "security": {
+            "should-verify-email": "authentication.security.should-verify-email"
+        },
+        "messaging": {
+            "controluser-registration-not-allowed-message": "authentication.messaging.controluser-registration-not-allowed-message",
+            "datauser-registration-not-allowed-message": "authentication.messaging.datauser-registration-not-allowed-message",
+            "already-registered": "authentication.messaging.already-registered"
+        }
+    },
+    "social-drivers": {
+        "github": {
+            "enabled": "social-drivers.github.enabled",
+            "client_id": "social-drivers.github.client_id",
+            "client_secret": "social-drivers.github.client_secret"
+        }
+    },
+    "control": {
+        "data-fields": {
+            "user": "control.data-fields.user",
+            "group": "control.data-fields.group",
+            "role": "control.data-fields.role",
+            "position": "control.data-fields.position"
+        }
+    },
+    "appearance": {
+        "messaging": {
+            "site-title": "appearance.messaging.site-title",
+            "footer": "appearance.messaging.footer",
+            "landing": {
+                "title": "appearance.messaging.landing.title"
+            }
+        }
+    }
+};